Message: Config project
Description: Added BaseFragment.kt
Added dagger

+++ app/src/main/java/com/utc/donlyconan/media/app/services/MusicalService.kt
@@ -23,14 +23,14 @@ class MusicalService: Service() {
 
     private val binder = object : IMusicalService.Stub() {
 
-        override fun setVideoId(videoId: Long) {
+        override fun setVideoId(videoId: Int) {
             Log.d(TAG, "setVideoId() called with: videoId = $videoId")
             this@MusicalService.videoId = videoId
         }
 
         override fun play() {
             Log.d(TAG, "play() called")
-            if(videoId != -1L) {
+            if(videoId != -1) {
                 val video = videoDao.getVideo(videoId)
                 Log.d(TAG, "play: video=$video")
                 player?.apply {
@@ -76,7 +76,7 @@ class MusicalService: Service() {
 
     }
 
-    var videoId: Long = -1L
+    var videoId: Int = -1
     private var player: MediaPlayer? = null
     @Inject lateinit var videoDao: VideoDao
