+++ app/src/main/java/com/utc/donlyconan/media/data/dao/VideoDao.kt
@@ -19,16 +19,16 @@ interface VideoDao {
     fun update(video: Video): Int
 
     @Query("Delete from videos where video_id = :videoId")
-    suspend fun delete(videoId: Long): Int
+    suspend fun delete(videoId: Int): Int
 
     @Query("Select * from videos where video_id = :videoId")
-    fun getVideo(videoId: Long): Video
+    fun getVideo(videoId: Int): Video
 
     @Query("Select * from videos where video_id > :fromId order by video_id asc limit 1")
-    fun getNextVideo(fromId: Long): Video
+    fun getNextVideo(fromId: Int): Video
 
     @Query("Select * from videos where video_id < :fromId order by video_id desc limit 1")
-    fun getPreviousVideo(fromId: Long): Video
+    fun getPreviousVideo(fromId: Int): Video
 
     @Query("Select count(video_id) from videos where path=:path")
     fun countPath(path: String): Int
